atlas parsers: fix 2 integer-overflow -> OOB on a malformed atlas (crash-safety)
A corrupted/truncated/partially-synced atlas .bin could wrap a uint32 arithmetic
PAST the bounds check -> OOB read -> WebContent crash on the fingerprint serve path:

(1) DriftstackWebGPUAtlas.mm: indexEntryStride is read as uint32 from the file
(unique among the atlas loaders -- the others use a constexpr size_t stride), so the
header check 'dataOffset != indexOffset + numEntries * indexEntryStride' and the
m_indexSpan subspan size were computed in 32-bit. numEntries=0x08000001, stride=32
wraps numEntries*stride to 0x20 -> passes the checks while m_numEntries stays ~134M
-> a 32-byte index span the later scan/binary-search overruns. Fix: compute indexSize
in uint64, add the explicit indexOffset+indexSize > fileSize bound, use it for the subspan.

(2) HTMLCanvasElement.cpp (V510 canvas fuzz atlas, the core toDataURL/toBlob/
OffscreenCanvas readback path, loaded in the production launch-env): per-entry
dataOff+dataLen (both uint32 from the file) were added in 32-bit, so dataOff=0xFFFFFF00,
dataLen=0x200 wraps to 0x100 -> passes '> payloadSize' -> subspan reads ~4GB past the
payload. Fix: uint64(dataOff)+uint64(dataLen) > size.

Behavior is identical for valid atlases (no wrap for in-range values) -- only a
malformed atlas (which would have crashed) is now rejected gracefully. Mirrors the
already-correct uint64 pattern in DriftstackAudioAtlas.mm / DriftstackEmojiAtlas.mm.
The other 7 atlas parsers audited robust. Build release exit 0; boots. Found by the
W2530 atlas-parser robustness audit.

+++ Source/WebCore/html/HTMLCanvasElement.cpp
@@ -1111,7 +1111,12 @@ static String v510AtlasLookupInState(const V510AtlasState& state, const String& 
                              | (uint32_t(entrySpan[18]) << 16) | (uint32_t(entrySpan[19]) << 24);
             uint32_t dataLen = uint32_t(entrySpan[20]) | (uint32_t(entrySpan[21]) << 8)
                              | (uint32_t(entrySpan[22]) << 16) | (uint32_t(entrySpan[23]) << 24);
-            if (dataOff + dataLen > state.dataPayloadSpan.size())
+            // W2530: compute dataOff+dataLen in 64-bit — both are uint32 read straight from the
+            // atlas file, so a 32-bit add WRAPS on a malformed/corrupt atlas (e.g. dataOff=0xFFFFFF00,
+            // dataLen=0x200 → 0x100), passing this bound while the subspan below reads ~4GB past the
+            // payload = WebContent crash. Promoting to uint64 makes the bound exact (cf. the already-
+            // correct DriftstackAudioAtlas.mm / DriftstackWebGPUAtlas.mm per-entry checks).
+            if (static_cast<uint64_t>(dataOff) + static_cast<uint64_t>(dataLen) > state.dataPayloadSpan.size())
                 return String();
             auto entry = state.dataPayloadSpan.subspan(dataOff, dataLen);
             static unsigned hits = 0;
